requirements.txt
ccxt
python-dotenv
customtkinter
pandas
matplotlib

# pip install -r requirements.txt
# python -m pip install pyyaml
